Create SubscriptionSelection component
